Fixed animation issues for sects
angular sections + their dividers with borders bearing the real size of the element creating layout issues. I decided to put these dividers in a wrapper and set the height of the wrapper to the effective height of the actual dividers. Then, I changed the height of the angular-content elements retroactively to account for the heights of these dividers. I will move these dividers outside of angular-section's in the future.